refactor(virrun): file-organization sweep over the rest of the exec services
The bwrap backend's two factory callbacks and the stderr status split
are model types instead of inline shapes. The untracked-entry hash, the
probe verdict grader and the gitignore entry writer leave the files
that read them for services of their own, and the task cache's age in
milliseconds sits beside the days it derives from.
